+++ sregistry/main/base/http.py
@@ -98,6 +98,21 @@ def paginate_get(self, url, headers=None, return_json=True, start_page=None):
     return results
         
 
+def verify(self):
+    '''
+       verify will return a True or False to determine to verify the
+       requests call or not. If False, we should the user a warning message,
+       as this should not be done in production!
+
+    '''
+    from sregistry.defaults import DISABLE_SSL_CHECK
+
+    if DISABLE_SSL_CHECK is True:
+        bot.warning('Verify of certificates disabled! ::TESTING USE ONLY::')
+
+    return not DISABLE_SSL_CHECK
+
+
 def download(self, url, file_name, headers=None, show_progress=True):
     '''stream to a temporary file, rename on successful completion
 
@@ -111,8 +126,11 @@ def download(self, url, file_name, headers=None, show_progress=True):
     fd, tmp_file = tempfile.mkstemp(prefix=("%s.tmp." % file_name)) 
     os.close(fd)
 
+    # Should we verify the request?
+    verify = self._verify()
+
     # Check here if exists
-    if requests.head(url).status_code in [200, 401]:
+    if requests.head(url, verify=verify).status_code in [200, 401]:
         response = self._stream(url,headers=headers,stream_to=tmp_file)
 
         if isinstance(response, HTTPError):
@@ -130,6 +148,8 @@ def stream(self, url, headers=None, stream_to=None, retry=True):
 
     bot.debug("GET %s" %url)
 
+
+
     # Ensure headers are present, update if not
     if headers == None:
         if self.headers is None:
@@ -138,6 +158,7 @@ def stream(self, url, headers=None, stream_to=None, retry=True):
 
     response = requests.get(url,         
                             headers=headers,
+                            verify=self._verify(),
                             stream=True)
 
     # Deal with token if necessary
@@ -204,6 +225,7 @@ def call(self, url, func, data=None, headers=None,
     response = func(url=url,
                     headers=heads,
                     data=data,
+                    verify=self._verify(),
                     stream=stream)
 
     # Errored response, try again with refresh