+++ sregistry/client/search.py
@@ -30,7 +30,7 @@ def main(args,parser,subparser):
             query = None
 
         try:
-            cli = get_client(query)
+            cli = get_client(query, args.quiet)
             cli.announce(args.command)
             cli.search(query=query, args=args)
         except NotImplementedError: