+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The versions coincide with releases on pip. Only major versions will be released
 ## [development](https://github.com/singularityware/singularity-python/tree/development) (development)
  - added Dropbox backend with push, pull, record, search, share (0.0.65) (development)
  - added ability for shell to take backend as argument (e.g., `sregistry shell dropbox`)
+ - added `--quiet` argument to client for option to suppression of client announcing itself.
  - shared function to get metadata from an image, checking for Singularity and inspecting if installed
  - removing sqlalchemy dependency, so user can optionally install database with client (0.0.64)
  - fixing bug with pull that not available to client for registry (0.0.63)